Refactoring of folders structure
