+++ src/blueprints/telegram_bot/_common/command_names.py
