+++ src/blueprints/telegram_bot/webhook/commands/__init__.py
@@ -22,7 +22,6 @@ not object directly
 """
 
 
-from .common.names import CommandName
 from .unknown import handle as unknown_handler
 from .help import handle as help_handler
 from .about import handle as about_handler
