+++ src/blueprints/telegram_bot/webhook/commands/settings.py
@@ -1,11 +1,11 @@
 from flask import g
 
 from src.api import telegram
-from .common.decorators import (
+from ._common.decorators import (
     register_guest,
     get_db_data
 )
-from .common.responses import (
+from ._common.responses import (
     request_private_chat
 )
