+++ src/blueprints/telegram_bot/webhook/commands/space.py
@@ -7,19 +7,19 @@ from plotly.express import colors
 from plotly.io import to_image
 
 from src.api import telegram
-from .common.responses import (
-    cancel_command,
-    AbortReason
-)
-from .common.decorators import (
-    yd_access_token_required,
-    get_db_data
-)
-from .common.yandex_api import (
+from src.blueprints.telegram_bot._common.yandex_disk import (
     get_disk_info,
     YandexAPIRequestError
 )
-from . import CommandName
+from ._common.responses import (
+    cancel_command,
+    AbortReason
+)
+from ._common.decorators import (
+    yd_access_token_required,
+    get_db_data
+)
+from src.blueprints.telegram_bot._common.command_names import CommandName
 
 
 @yd_access_token_required
