+++ src/blueprints/telegram_bot/webhook/commands/unknown.py
@@ -1,7 +1,7 @@
 from flask import g
 
 from src.api import telegram
-from . import CommandName
+from src.blueprints.telegram_bot._common.command_names import CommandName
 
 
 def handle(*args, **kwargs):
